# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auditoria": {
      "post": {
        "operationId": "AuditoriaController_post",
        "summary": "Crear un nueva aditoria",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "La aditoria ha sido creada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "auditoria"
        ]
      },
      "get": {
        "operationId": "AuditoriaController_getAll",
        "summary": "Obtener todas las aditorias",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": " query-Filter.e.g.col1:v1,col2:v2… ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve todas las aditorias.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditoriaDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auditoria"
        ]
      }
    },
    "/auditoria/auditor/{personaId}": {
      "get": {
        "operationId": "AuditoriaController_getByAuditor",
        "summary": "Obtener auditorías por auditor",
        "parameters": [
          {
            "name": "personaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": " query-Filter.e.g.col1:v1,col2:v2… ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve las auditorías del auditor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditoriaDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auditoria"
        ]
      }
    },
    "/auditoria/{id}": {
      "get": {
        "operationId": "AuditoriaController_getById",
        "summary": "Obtener una auditoria por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve la auditoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDTO"
                }
              }
            }
          },
          "404": {
            "description": "Auditoria no encontrada."
          }
        },
        "tags": [
          "auditoria"
        ]
      },
      "put": {
        "operationId": "AuditoriaController_put",
        "summary": "Actualizar una auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La auditoria ha sido actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Contratista no encontrado."
          }
        },
        "tags": [
          "auditoria"
        ]
      },
      "delete": {
        "operationId": "AuditoriaController_delete",
        "summary": "Eliminar una auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La auditoria ha sido eliminada exitosamente."
          },
          "404": {
            "description": "Auditoria no encontrada."
          }
        },
        "tags": [
          "auditoria"
        ]
      }
    },
...
    "/auditor/{id}": {
      "get": {
        "operationId": "AuditorController_getById",
        "summary": "Obtener un auditor por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve la auditor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditorDTO"
                }
              }
            }
          },
          "404": {
            "description": "Auditor no encontrada."
          }
        },
        "tags": [
          "auditor"
        ]
      },
      "put": {
        "operationId": "AuditorController_put",
        "summary": "Actualizar un auditor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditorDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La auditor ha sido actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditorDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Auditor no encontrado."
          }
        },
        "tags": [
          "auditor"
        ]
      },
      "delete": {
        "operationId": "AuditorController_delete",
        "summary": "Eliminar un auditor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La auditor ha sido eliminada exitosamente."
          },
          "404": {
            "description": "Auditor no encontrada."
          }
        },
        "tags": [
          "auditor"
        ]
      }
    },
...
    "/auditoria-gestion": {
      "post": {
        "operationId": "AuditoriaGestionController_post",
        "summary": "Crear una nueva auditoria junto con su estado inicial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditoriaGestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "La auditoria ha sido creada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAuditoriaGestionDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "auditoria-gestion"
        ]
      }
    },
    "/auditoria-gestion/{id}": {
      "put": {
        "operationId": "AuditoriaGestionController_put",
        "summary": "Actualizar los estados de las auditorias existentes en un plan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del plan de auditorias a actualizar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriaEstadoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Los estados de las auditorias han sido actualizados exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaEstadoDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Plan de Auditorias no encontrado."
          }
        },
        "tags": [
          "auditoria-gestion"
        ]
      }
    }
  },
  "info": {
    "title": "Plan AnualAuditoria CRUD",
    "description": "API CRUD para la gestión de las auditoría y planes de mejoramiento",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AuditoriaDTO": {
        "type": "object",
        "properties": {
          "plan_auditoria_id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "tipo_evaluacion_id": {
            "type": "number"
          },
          "cronograma_id": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado_id": {
            "type": "number"
          },
          "no_auditoria": {
            "type": "number"
          },
          "vigencia_id": {
            "type": "number"
          },
          "consecutivo_OCI": {
            "type": "string"
          },
          "consecutivo_IE": {
            "type": "string"
          },
          "tipo_id": {
            "type": "number"
          },
          "macroproceso": {
            "type": "number"
          },
          "macroproceso_id": {
            "type": "number"
          },
          "proceso_id": {
            "type": "number"
          },
          "dependencia_id": {
            "type": "number"
          },
          "lider_id": {
            "type": "number"
          },
          "responsable_id": {
            "type": "number"
          },
          "fecha_inicio": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_fin": {
            "format": "date-time",
            "type": "string"
          },
          "objetivo": {
            "type": "string"
          },
          "alcance": {
            "type": "string"
          },
          "criterio": {
            "type": "string"
          },
          "rec_tecnologico": {
            "type": "string"
          },
          "rec_humano": {
            "type": "string"
          },
          "rec_fisico": {
            "type": "string"
          },
          "temas": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "plan_auditoria_id",
          "titulo",
          "tipo_evaluacion_id",
          "cronograma_id",
          "estado_id",
          "no_auditoria",
          "vigencia_id",
          "consecutivo_OCI",
          "consecutivo_IE",
          "tipo_id",
          "macroproceso",
          "macroproceso_id",
          "proceso_id",
          "dependencia_id",
          "lider_id",
          "responsable_id",
          "fecha_inicio",
          "fecha_fin",
          "objetivo",
          "alcance",
          "criterio",
          "rec_tecnologico",
          "rec_humano",
          "rec_fisico",
          "temas",
          "activo",
          "fecha_creacion",
          "fecha_modificacion"
        ]
      },
...
      "AuditoriaEstadoDto": {
        "type": "object",
        "properties": {
          "auditoria_id": {
            "type": "string"
          },
          "usuario_id": {
            "type": "number"
          },
          "usuario_rol": {
            "type": "string"
          },
          "observacion": {
            "type": "string"
          },
          "actual": {
            "type": "boolean"
          },
          "estado_id": {
            "type": "number"
          },
          "fase_id": {
            "type": "string"
          },
          "fecha_ejecucion_estado": {
            "format": "date-time",
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        },
        "required": [
          "auditoria_id",
          "usuario_id",
          "usuario_rol",
          "observacion",
          "actual",
          "estado_id",
          "fase_id",
          "fecha_ejecucion_estado",
          "activo"
        ]
      },
...
      "AuditorDTO": {
        "type": "object",
        "properties": {
          "auditoria_id": {
            "type": "string"
          },
          "auditor_id": {
            "type": "number"
          },
          "asignado": {
            "type": "boolean"
          },
          "asignado_por_id": {
            "type": "number"
          },
          "auditor_lider": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "auditoria_id",
          "auditor_id",
          "asignado",
          "asignado_por_id",
          "auditor_lider",
          "activo",
          "fecha_creacion",
          "fecha_modificacion"
        ]
      },
...
      "CreateAuditoriaGestionDto": {
        "type": "object",
        "properties": {
          "plan_auditoria_id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "tipo_evaluacion_id": {
            "type": "number"
          },
          "cronograma_id": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado_id": {
            "type": "number"
          },
          "no_auditoria": {
            "type": "number"
          },
          "vigencia_id": {
            "type": "number"
          },
          "consecutivo_OCI": {
            "type": "string"
          },
          "consecutivo_IE": {
            "type": "string"
          },
          "tipo_id": {
            "type": "number"
          },
          "macroproceso": {
            "type": "number"
          },
          "macroproceso_id": {
            "type": "number"
          },
          "proceso_id": {
            "type": "number"
          },
          "dependencia_id": {
            "type": "number"
          },
          "lider_id": {
            "type": "number"
          },
          "responsable_id": {
            "type": "number"
          },
          "fecha_inicio": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_fin": {
            "format": "date-time",
            "type": "string"
          },
          "objetivo": {
            "type": "string"
          },
          "alcance": {
            "type": "string"
          },
          "criterio": {
            "type": "string"
          },
          "rec_tecnologico": {
            "type": "string"
          },
          "rec_humano": {
            "type": "string"
          },
          "rec_fisico": {
            "type": "string"
          },
          "temas": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          },
          "auditoria_id": {
            "type": "string"
          },
          "usuario_id": {
            "type": "number"
          },
          "usuario_rol": {
            "type": "string"
          },
          "observacion": {
            "type": "string"
          },
          "actual": {
            "type": "boolean"
          },
          "fase_id": {
            "type": "string"
          },
          "fecha_ejecucion_estado": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "plan_auditoria_id",
          "titulo",
          "tipo_evaluacion_id",
          "cronograma_id",
          "estado_id",
          "no_auditoria",
          "vigencia_id",
          "consecutivo_OCI",
          "consecutivo_IE",
          "tipo_id",
          "macroproceso",
          "macroproceso_id",
          "proceso_id",
          "dependencia_id",
          "lider_id",
          "responsable_id",
          "fecha_inicio",
          "fecha_fin",
          "objetivo",
          "alcance",
          "criterio",
          "rec_tecnologico",
          "rec_humano",
          "rec_fisico",
          "temas",
          "activo",
          "fecha_creacion",
          "fecha_modificacion",
          "auditoria_id",
          "usuario_id",
          "usuario_rol",
          "observacion",
          "actual",
          "fase_id",
          "fecha_ejecucion_estado"
        ]
      }
    }
  }
}